+++ src/main/java/com/proyecto/core/stock/application/dto/StockPorSedeDto.java
@@ -1,7 +1,7 @@
 package com.proyecto.core.stock.application.dto;
 
 public record StockPorSedeDto(
-    Integer idSede,
+    Long idSede,
     String nombre,
     Integer stockTotal
 ) {}
